FEAT: Add default scroll layout
